move CloudProvider enum to its own crate
