+++ Cargo.toml
@@ -17,6 +17,7 @@ members = [
     "src/catalog-debug",
     "src/catalog-protos",
     "src/cloud-api",
+    "src/cloud-provider",
     "src/cluster",
     "src/clusterd",
     "src/cluster-client",
@@ -134,6 +135,7 @@ default-members = [
     "src/catalog-protos",
     "src/ccsr",
     "src/cloud-api",
+    "src/cloud-provider",
     "src/cloud-resources",
     "src/cluster",
     "src/cluster-client",
