+++ src/adapter/Cargo.toml
@@ -37,6 +37,7 @@ mz-audit-log = { path = "../audit-log" }
 mz-build-info = { path = "../build-info" }
 mz-catalog = { path = "../catalog" }
 mz-ccsr = { path = "../ccsr" }
+mz-cloud-provider = { path = "../cloud-provider", default-features = false }
 mz-cloud-resources = { path = "../cloud-resources" }
 mz-cluster-client = { path = "../cluster-client" }
 mz-compute-client = { path = "../compute-client" }
