+++ src/adapter/src/config/frontend.rs
@@ -14,8 +14,9 @@ use std::time::Duration;
 use derivative::Derivative;
 use launchdarkly_server_sdk as ld;
 use mz_build_info::BuildInfo;
+use mz_cloud_provider::CloudProvider;
 use mz_ore::now::NowFn;
-use mz_sql::catalog::{CloudProvider, EnvironmentId};
+use mz_sql::catalog::EnvironmentId;
 use tokio::time;
 
 use crate::config::{Metrics, SynchronizedParameters, SystemParameterSyncConfig};
