+++ src/cloud-provider/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "mz-cloud-provider"
+description = "Representation of a cloud provider"
+version = "0.1.0"
+authors = ["Materialize, Inc."]
+license = "proprietary"
+edition.workspace = true
+rust-version.workspace = true
+publish = false
+
+[dependencies]
+workspace-hack = { version = "0.0.0", path = "../workspace-hack", optional = true }
+
+[features]
+default = ["workspace-hack"]
+
+[lints]
+workspace = true
