+++ src/orchestratord/src/controller/materialize.rs
@@ -21,13 +21,13 @@ use serde::Deserialize;
 use tracing::{debug, trace};
 
 use crate::metrics::Metrics;
+use mz_cloud_provider::CloudProvider;
 use mz_cloud_resources::crd::materialize::v1alpha1::{
     Materialize, MaterializeCertSpec, MaterializeStatus,
 };
 use mz_orchestrator_kubernetes::KubernetesImagePullPolicy;
 use mz_orchestrator_tracing::TracingCliArgs;
 use mz_ore::{cast::CastFrom, cli::KeyValueArg, instrument};
-use mz_sql::catalog::CloudProvider;
 
 pub mod balancer;
 pub mod console;
